added the 'stock' firmware binary file
obtained by the simple compilation, note that the bootloader in the AR9271 does not do any Integrity test on the firmware.
Note also that the firmware, before being converted to a binary file, is a 700kB approx. ELF, this may comport some complication when it will be processed in the memory process.